backend/src/food_recipes/requirements.txt
Django==4.2
django-autoslug==1.9.9
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
